pyproject.toml
[tool.poetry]
name = "iubh-computer-vision"
version = "1.0.0"
description = "IUBH Computer Vision project, evalulate cutting edge of open-source computer vision models"
authors = ["Sulaiman Ghori <sulaiman.ghori@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
transformers = "^4.29.2"
pillow = "^9.5.0"
timm = "^0.9.2"
matplotlib = "^3.7.1"
scipy = "^1.11.1"
pydantic = "^2.0.3"
datasets = {version = "^2.14.0", extras = ["vision"]}
torch = { url = "https://download.pytorch.org/whl/nightly/cpu/torch-2.1.0.dev20230725-cp311-none-macosx_11_0_arm64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/nightly/cpu/torchvision-0.16.0.dev20230725-cp311-cp311-macosx_11_0_arm64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/nightly/cpu/torchaudio-2.1.0.dev20230725-cp311-cp311-macosx_11_0_arm64.whl" }
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.21.1"
opencv-python = "^4.8.0.74"
pandas = "^2.0.3"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
